requirements.txt
lxml==4.4.1
Pillow==6.2.0
python-pptx==0.6.18
XlsxWriter==1.2.1
